+++ src/Calendar/docs/index.story.js
@@ -82,6 +82,36 @@ export default {
       'cs',
       'uk',
     ],
+    firstDayOfWeek: [
+      {
+        label: `0`,
+        value: 0,
+      },
+      {
+        label: `1`,
+        value: 1,
+      },
+      {
+        label: `2`,
+        value: 2,
+      },
+      {
+        label: `3`,
+        value: 3,
+      },
+      {
+        label: `4`,
+        value: 4,
+      },
+      {
+        label: `5`,
+        value: 5,
+      },
+      {
+        label: `6`,
+        value: 6,
+      },
+    ],
   },
 
   sections: [