+++ src/DatePicker/test/DatePicker.spec.js
@@ -675,6 +675,21 @@ describe('DatePicker', () => {
         new Date('2017-01-02T12:34:56.000Z'),
       );
     });
+
+    it('should adjust time if no value given to midnight', () => {
+      const onChange = jest.fn();
+      const { calendarDriver, driver } = createDriver(
+        <DatePicker onChange={onChange} />,
+      );
+      driver.open();
+      calendarDriver.clickOnNthDay(1);
+
+      const chosenDate = onChange.mock.calls[0][0];
+      expect(chosenDate.getHours()).toEqual(0);
+      expect(chosenDate.getMinutes()).toEqual(0);
+      expect(chosenDate.getSeconds()).toEqual(0);
+      expect(chosenDate.getMilliseconds()).toEqual(0);
+    });
   });
 
   describe('`readonly` prop', () => {
@@ -824,6 +839,18 @@ describe('DatePicker', () => {
       );
       inputDriver.trigger('click');
       expect(calendarDriver.isTwoMonthsLayout()).toBe(true);
+    });
+  });
+
+  describe('firstDayOfWeek', () => {
+    it('should show correct first day of the week', async () => {
+      const { inputDriver, calendarDriver } = createDriver(
+        <DatePicker onChange={() => {}} firstDayOfWeek={0} />,
+      );
+
+      inputDriver.trigger('click');
+
+      expect(await calendarDriver.getNthWeekDayName(0)).toEqual('Su');
     });
   });
 });