+++ src/FacesRatingBar/FacesRatingBar.meta.js
@@ -1,0 +1,8 @@
+import FacesRatingBar from './FacesRatingBar';
+import Registry from '@ui-autotools/registry';
+
+const metadata = Registry.getComponentMetadata(FacesRatingBar);
+
+metadata.exportedFrom({
+  path: 'src/FacesRatingBar/FacesRatingBar.js',
+});