+++ src/TagList/index.d.ts
@@ -1,17 +1,20 @@
 import * as React from 'react';
 import { TagProps } from '../Tag';
 import { ButtonProps } from '../Button';
+import { OmitPolyfill } from '../common';
+
+export type TagListSize = 'small' | 'medium' | 'large';
 
 export type TagListProps = {
   dataHook?: string;
-  tags: Array<TagProps>;
+  tags: Array<OmitPolyfill<TagProps, 'size' | 'onRemove'>>;
+  size?: TagListSize;
+  onTagRemove?: (id: string) => void;
   actionButton?: TagListActionButtonProps;
-}
+};
 
-type TagListActionButtonProps = ButtonProps & {
+type TagListActionButtonProps = Pick<ButtonProps, 'onClick'> & {
   label?: string;
-}
+};
 
-export default class TagList extends React.PureComponent<TagListProps>{
-  
-}
+export default class TagList extends React.PureComponent<TagListProps> {}