+++ stories/index.js
@@ -53,6 +53,7 @@ require('../src/EditableSelector/docs/index.story');
 require('../src/EditableTitle/docs/index.story');
 require('../src/EmptyState/docs/index.story');
 require('../src/EndorseContentLayout/docs');
+require('../src/FacesRatingBar/docs/index.story');
 require('../src/FilePicker/docs/index.story');
 require('../src/FileUpload/docs/index.story');
 require('../src/FillButton/docs/index.story');